+++ src/main/java/bridge/BridgeRandomNumberGenerator.java
@@ -1,5 +1,6 @@
-package bridge.domain.generator;
+package bridge;
 
+import bridge.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class BridgeRandomNumberGenerator implements BridgeNumberGenerator {
