+++ src/test/java/bridge/BridgeRandomNumberGeneratorTest.java
@@ -1,6 +1,5 @@
 package bridge;
 
-import bridge.domain.generator.BridgeRandomNumberGenerator;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
