+++ src/test/java/bridge/domain/BridgeMakerTest.java
@@ -1,6 +1,6 @@
 package bridge.domain;
 
-import bridge.domain.generator.BridgeRandomNumberGenerator;
+import bridge.BridgeRandomNumberGenerator;
 import bridge.domain.generator.BridgeMaker;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
